Updated comparisonplotter.py with clarity, and a data fix in C# for the split part

+++ StudyPython/StudyPython.py
@@ -234,8 +234,8 @@ def split_train_infer(
 
     # --- Split phase ---
     def _split():
-        X = sub[features].to_numpy(dtype=np.float32, copy=False)
-        y = sub[TARGET].to_numpy(dtype=np.float32, copy=False)
+        X = sub[features].to_numpy(dtype=np.float64, copy=False)
+        y = sub[TARGET].to_numpy(dtype=np.float64, copy=False)
         return train_test_split(X, y, test_size=test_size, random_state=seed)
 
     (X_train, X_test, y_train, y_test), split_ns, split_heap, split_rss = _measure(_split)
